NuGet packaging.
Switched MyMediaLite itself to use NuGet for dependency
management. The versions of C5, MathNet.Numerics, and Nunit
used were old and don't have a package on nuget.org so I
upgraded to the latest versions of them.

Fixed warnings in the unit tests caused by using an NUnit
method that is now marked as obsolete.

Wrote a DelegateComparer class in MyMediaLite because it is
not present in newer C5 versions.

Wrote a .nuspec for MyMediaLite. Zeno should look it over and
expand on the description at least.

Wrote a BuildPackage MSBuild target.